requirements.txt
pillow
numpy
matplotlib
scikit-image
bresenham
